splitted credits and casino commands
