+++ Doug/Startup.cs
@@ -56,6 +56,7 @@ namespace Doug
             services.AddTransient<ICoffeeCommands, CoffeeCommands>();
             services.AddTransient<ISlursCommands, SlursCommands>();
             services.AddTransient<ICreditsCommands, CreditsCommands>();
+            services.AddTransient<ICasinoCommands, CasinoCommands>();
             services.AddTransient<ICoffeeService, CoffeeService>();
 
             services.AddTransient<IChannelRepository, ChannelRepository>();
